requirements.txt
flask
gunicorn
google-generativeai
supabase
python-dotenv
pydantic
json